LayoutTests/imported/w3c/web-platform-tests/css/css-flexbox/alignment/flex-align-baseline-006-expected.txt





PASS #target > div 1
FAIL #target > div 2 assert_equals:
<div data-offset-y="0"><span></span><br><span></span></div>
offsetTop expected 0 but got 20
FAIL #target > div 3 assert_equals:
<div data-offset-y="55"><span></span><br><span></span></div>
offsetTop expected 55 but got 65
